# docs: explain why TUI is reinitialized on resume 