Every time the TUI is suspended and resumed we completely reinitialize the TUI's terminfo state: we re-parse the terminfo database file, patch terminfo according to the terminal type, query the terminal for supported capabilities, etc. This is wasteful because in the overwhelming majority of cases nothing about the terminal has changed at all between suspend and resume.

Instead, initialize terminfo and query the terminal only once when Nvim is started and maintain this state between suspends and resumes.

---

Benchmark measuring time to resume after a suspend in milliseconds taken from my 2021 M1 MacBook Pro in a Debug build:

| **master** | **PR** |
| ---------- | ------ |
| 4.679      | 3.272  |
| 4.982      | 3.447  |
| 5.280      | 2.919  |
| 4.697      | 2.974  |

The improvement in milliseconds is not very dramatic (5 ms is well below the human perception threshold), so this is more about simplification than about performance (it also eliminates some unnecessary I/O to the terminal since we are no longer querying for e.g. kitty keyboard or synchronized output support on every resume).