@zeertzjq 2024-01-25
> This may be useful for some multiplexers which are able to forward these sequences to the parent terminal.

Ah yes, that's not uncommon.