@zeertzjq 2024-01-25
I just remembered this: https://github.com/nelhage/reptyr