@gpanders 2024-01-25
> I just remembered this: https://github.com/nelhage/reptyr

Interesting, in this case even `$TERM` could (legitimately) change.

I think I'm leaning towards closing this and leaving the status quo as-is. It's bothersome in an abstract sense that we are doing a bunch of wasted work, but practically speaking that work does not take very much time and has basically no impact on the user experience. And given that there are some (albeit, rare) use cases, it does not seem worth it.