@justinmk 2024-01-25
The docstrings in this PR are an improvement, perhaps also worth mentioning  https://github.com/nelhage/reptyr in the docstring to explain why we "reprocess terminfo and requery the terminal" after suspend/resume.

> bothersome in an abstract sense that we are doing a bunch of wasted work, but practically speaking that work does not take very much time

Eliminating "noise" does help with troubleshooting and debugging. But this does seem like a case where "correct" is less confusing that depending on state that happens to _usually_ not change.